requirements.txt
flask==2.2.3
werkzeug==2.2.3
pyodbc==5.2.0
gunicorn==20.1.0
python-dotenv==1.0.0